requirements.txt
pyyaml==6.0.2
plotly==6.1.2
numpy==1.25.0
open3d==0.19.0 
Pillow==11.0.0 
PyQt5
vtk
opencv-python
pyqt5-qt5
